requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gevent==23.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
